requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
coloredlogs==15.0.1
Django==5.2.8
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-js-asset==3.1.2
django-modeltranslation==0.19.17
humanfriendly==10.0
idna==3.11
npyscreen==4.10.5
pillow==12.0.0
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.3
tabulate==0.9.0
tcore==0.2.3
typing_extensions==4.15.0
urllib3==2.5.0
